requirements.txt
feedparser==6.0.11
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.11.0
